Add view revision notes feature
This commit adds the ability to view revision notes for a data item in the dataTable component. It also adjusts the max-width of the verification/revision dialog based on the type of dialog being shown. In addition, it refactors the dialog component's computed property to a watch function and resets revisionNotes field upon dialog close.