Refactor components used for dialogs
This commit refactors the implementation of Vue components used in dialogs by creating a new generic component "appDialog.vue" that replaces the previous specific components. "appDialog.vue" uses dynamic components, making it more flexible for displaying different types of dialogs. Also, it sets the width of the dialog based on the type of dialog being displayed. Additionally, the commit includes the creation of new components "revisionDataDialog.vue" and "revisionNoteListDialog.vue". These components are used for displaying and revising revision notes for a specific data and listing the revision history of a specific data, respectively.